fix(web): prevent clearing fallback header
If the cookie fallback isn't set in local storage, the X-Fallback-Cookies header is set to empty even if the default header has some value set. This change will only set the header if there's a value in local storage to allow developer to set the fallback header if they have a value from somewhere else.