fix(cli): use JSONbig to parse and stringify JSON
Due to the way JavaScript handles numbers, the standard JSON parser
will convert the max PHP int, 9223372036854775807, to
9223372036854776000. However, this is problematic because the resulting
number is an invalid int in PHP so the integer validation fails.

This PR ensures we use JSONbig to parse and stringify JSON in the CLI
to retain the precision of the numbers.